requirements.txt
imageio==2.25.0
Pillow==9.4.0
